EM-1911 Move services into services folder
