+++ projects/media-viewer/src/lib/annotations/annotation-set/annotation-set.model.ts
@@ -1,4 +1,4 @@
-import { ApiPersisted } from '../api-persisted.model';
+import { ApiPersisted } from '../models/api-persisted.model';
 import { Annotation } from './annotation-view/annotation.model';
 
 export interface AnnotationSet extends ApiPersisted {
