+++ projects/media-viewer/src/lib/annotations/comment-set/comment/comment.component.spec.ts
@@ -6,7 +6,7 @@ import { CommentService } from './comment.service';
 import { TextHighlightDirective } from './text-highlight.directive';
 import { TagsServices } from '../../services/tags/tags.services';
 import { HttpClientTestingModule } from '@angular/common/http/testing';
-import { MomentDatePipe } from '../../pipes/date.pipe';
+import { MomentDatePipe } from '../../pipes/date/date.pipe';
 import { StoreModule } from '@ngrx/store';
 import { reducers } from '../../../store/reducers/reducers';
