+++ projects/media-viewer/src/lib/annotations/comment-set/comment/comment.service.spec.ts
@@ -5,7 +5,7 @@ import { ComponentFixture, fakeAsync, TestBed } from '@angular/core/testing';
 import { NO_ERRORS_SCHEMA } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 import { HttpClientTestingModule } from '@angular/common/http/testing';
-import { AnnotationApiService } from '../../annotation-api.service';
+import { AnnotationApiService } from '../../services/annotation-api/annotation-api.service';
 import { ToolbarEventService } from '../../../toolbar/toolbar-event.service';
 import { CommentSetRenderService } from '../comment-set-render.service';
 import { annotationSet } from '../../../../assets/annotation-set';
