+++ projects/media-viewer/src/lib/annotations/models/api-persisted.model.ts
@@ -1,4 +1,4 @@
-import { User } from './models/user.model';
+import { User } from './user.model';
 
 export interface ApiPersisted {
   id: string;
