+++ projects/media-viewer/src/lib/annotations/pipes/filter/filter.pipe.ts
