+++ projects/media-viewer/src/lib/annotations/services/annotation-api/annotation-api.service.spec.ts
@@ -1,11 +1,12 @@
 import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
 import { fakeAsync, inject, TestBed } from '@angular/core/testing';
-import { AnnotationApiService } from './annotation-api.service';
-import { AnnotationSet } from './annotation-set/annotation-set.model';
-import { User } from './models/user.model';
-import { Annotation } from './annotation-set/annotation-view/annotation.model';
 import { of } from 'rxjs';
-import { Comment } from './comment-set/comment/comment.model';
+
+import { AnnotationApiService } from './annotation-api.service';
+import { AnnotationSet } from '../../annotation-set/annotation-set.model';
+import { User } from '../../models/user.model';
+import { Annotation } from '../../annotation-set/annotation-view/annotation.model';
+import { Comment } from '../../comment-set/comment/comment.model';
 
 describe('AnnotationApiService', () => {
   let httpMock: HttpTestingController;
