+++ projects/media-viewer/src/lib/annotations/services/annotation-api/annotation-api.service.ts
@@ -1,11 +1,12 @@
-import { Observable } from 'rxjs';
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
+import { Observable } from 'rxjs';
 import {catchError, map} from 'rxjs/operators';
 import uuid from 'uuid/v4';
-import { AnnotationSet } from './annotation-set/annotation-set.model';
-import { Annotation } from './annotation-set/annotation-view/annotation.model';
-import { Comment } from './comment-set/comment/comment.model';
+
+import { AnnotationSet } from '../../annotation-set/annotation-set.model';
+import { Annotation } from '../../annotation-set/annotation-view/annotation.model';
+import { Comment } from '../../comment-set/comment/comment.model';
 
 @Injectable()
 export class AnnotationApiService {
