+++ projects/media-viewer/src/lib/store/effects/annotations.effects.spec.ts
@@ -4,7 +4,7 @@ import { hot, cold } from 'jasmine-marbles';
 import { of } from 'rxjs';
 import { provideMockActions } from '@ngrx/effects/testing';
 import * as annotationActions from '../actions/annotations.action';
-import {AnnotationApiService} from '../../annotations/annotation-api.service';
+import {AnnotationApiService} from '../../annotations/services/annotation-api/annotation-api.service';
 import {AnnotationEffects} from './annotations.effect';
