+++ projects/media-viewer/src/lib/store/effects/bookmarks.effects.spec.ts
@@ -4,7 +4,7 @@ import { of, throwError } from 'rxjs';
 import { provideMockActions } from '@ngrx/effects/testing';
 import * as bookmarkActions from '../actions/bookmarks.action';
 import { BookmarksEffects } from './bookmarks.effects';
-import { BookmarksApiService } from '../../annotations/bookmarks-api.service';
+import { BookmarksApiService } from '../../annotations/services/bookmarks-api/bookmarks-api.service';
 import { Store, StoreModule } from '@ngrx/store';
 import { reducers } from '../reducers/reducers';
 import { PdfPositionUpdate } from '../actions/document.action';
@@ -143,4 +143,3 @@ describe('Bookmark Effects', () => {
     });
   });
 });
-
